+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 * Added Azure Container Apps resource detector.
   ([#41803](https://github.com/Azure/azure-sdk-for-net/pull/41803))
+* Added `Azure.Monitor.OpenTelemetry.LiveMetrics`, enabling the sending of [live
+  metrics
+  data](https://learn.microsoft.com/azure/azure-monitor/app/live-stream).
+  The newly added `EnableLiveMetrics` property is set to `true` by default. This
+  property can be set to `false` to disable live metrics.
+  ([#41872](https://github.com/Azure/azure-sdk-for-net/pull/41872))
 
 ### Breaking Changes